fix. 리뷰반영 - balance, amount를 long 타입으로 변경 by 양성빈

+++ jaehoon/java-mid1/src/main/java/section06/Atm.java
@@ -4,7 +4,7 @@ import static java.util.Objects.isNull;
 
 public class Atm {
 
-  public TransactionResult processTransaction(Member member, Member targetMember, String operation, double amount) {
+  public TransactionResult processTransaction(Member member, Member targetMember, String operation, long amount) {
     TransactionType transactionType = TransactionType.findByTransactionType(operation);
     if (isNull(transactionType)) {
       throw new IllegalArgumentException("지원되지 않는 거래 유형입니다: " + operation);
@@ -17,19 +17,19 @@ public class Atm {
     };
   }
 
-  private TransactionResult processDeposit(Member member, double amount) {
+  private TransactionResult processDeposit(Member member, long amount) {
     BankAccount updatedAccount = member.getBankAccount().deposit(amount);
     member.updateBankAccount(updatedAccount);
     return new TransactionResult(TransactionType.DEPOSIT, amount, member, null);
   }
 
-  private TransactionResult processWithdraw(Member member, double amount) {
+  private TransactionResult processWithdraw(Member member, long amount) {
     BankAccount updatedAccount = member.getBankAccount().withdraw(amount);
     member.updateBankAccount(updatedAccount);
     return new TransactionResult(TransactionType.WITHDRAW, amount, member, null);
   }
 
-  private TransactionResult processTransfer(Member member, Member targetMember, double amount) {
+  private TransactionResult processTransfer(Member member, Member targetMember, long amount) {
     if (targetMember == null || member.equals(targetMember)) {
       throw new IllegalArgumentException("송금 대상자가 유효하지 않습니다.");
     }
